# Speed up basic quantity operations.