By removing duplication, this ensures that, e.g., `a.to_value(u.radian)` is not slower than a.radian.  It also speeds up binary operations where no unit change is necessary, and thus things like `np.add.reduce(a)`.

This got started while about to tell @adrn to use `to_value(u.radian)` rather than `.radian`, only to notice that the latter was faster. Which it really has no reason to be. Furthermore, in principle, overrides of `sum` and so on, which just call `reduce` operations on ufuncs, should not be necessary any more with numpy 1.13; however, the reductions were much slower. So, those got optimized as well (i.e., I got carried away with doing stuff for 3.1 rather than reviewing stuff for 3.0...)

Timing improvements:
```
import numpy as np, astropy.units as u
from astropy.coordinates import Angle
a = Angle(np.arange(10.), 'degree')
%timeit a.radian
# 10000 loops, best of 3: 25.5 µs per loop  --> no change
%timeit a.to_value(u.radian)
# 10000 loops, best of 3: 33.4 µs per loop  --> 16.5 µs
%timeit a+a
# 10000 loops, best of 3: 86.5 µs per loop  --> 77.3 µs
%timeit np.add.reduce(a)
# 10000 loops, best of 3: 114 µs per loop  --> 78.2 µs
```