@mhvk 2017-12-21
Thanks, @adrn; I went ahead and merged since there was no reason not so.